requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
aiofiles
sqlalchemy
passlib
psycopg2-binary
python-multipart
python-jose
bcrypt
starlette-admin
sqlalchemy-file
Pillow
fasteners